override SubmitBeaconBlock()
